+++ src/components/loader-bar/loader-bar.style.js
@@ -24,20 +24,20 @@ const innerBarAnimation = keyframes`
 `;
 
 const StyledLoaderBar = styled.div`
-  ${({ size, theme }) => css`
+  ${({ size }) => css`
     display: inline-block;
     height: ${getHeight(size)};
     width: 100%;
-    background-color: ${theme.colors.loadingBarBackground};
+    background-color: var(--colorsActionMajor150);
     overflow: hidden;
     position: relative;
   `}
 `;
 
 const InnerBar = styled.div`
-  ${({ theme, size }) => css`
+  ${({ size }) => css`
     position: absolute;
-    background-color: ${theme.colors.primary};
+    background-color: var(--colorsActionMajor500);
     width: ${INNER_BAR_LENGTH};
     height: ${getHeight(size)};
     animation: 2s ${innerBarAnimation} linear 0s infinite normal none running;
@@ -60,12 +60,10 @@ StyledLoader.defaultProps = {
 };
 
 StyledLoaderBar.defaultProps = {
-  theme: baseTheme,
   size: "medium",
 };
 
 InnerBar.defaultProps = {
-  theme: baseTheme,
   size: "medium",
 };